feat udistrital/arka_cliente#1111: Consecutivo movimientos
- Refactor módulo de bajas.
